refactor: Move all type conversions into `conversions.ts`
Attempting to clean up imports a bit
