+++ src/components/Interactive.tsx
@@ -1,4 +1,5 @@
 import React, { useState, useLayoutEffect, useRef, useCallback } from "react";
+
 import styles from "../styles.css";
 
 // Limit number within [0, 1] bounds.
