+++ src/packages/rgb.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import ColorPicker from "../components/ColorPicker";
 
-import toHsv from "../utils/rgbToHsv";
-import fromHsv from "../utils/hsvToRgb";
-import equal from "../utils/equalColorObjects";
+import ColorPicker from "../components/ColorPicker";
 import { BaseComponentProps, RGB } from "../types";
+import { rgbToHsv, hsvToRgb } from "../utils/conversions";
+import equal from "../utils/equalColorObjects";
 
 interface Props extends BaseComponentProps {
   className: string;
@@ -18,8 +17,8 @@ const Rgb: React.FC<Props> = (props) => {
       className={props.className}
       colorModel={{
         defaultColor: { r: 0, g: 0, b: 0 },
-        toHsv,
-        fromHsv,
+        toHsv: rgbToHsv,
+        fromHsv: hsvToRgb,
         equal,
       }}
       color={props.color}
