--- src/utils/hsvToHslString.ts
@@ -1,9 +1,0 @@
-import hsvToHsl from "./hsvToHsl";
-import { HSV } from "../types";
-
-const hsvToHslString = (hsv: HSV): string => {
-  const { h, s, l } = hsvToHsl(hsv);
-  return `hsl(${h}, ${s}%, ${l}%)`;
-};
-
-export default hsvToHslString;
